started calender 2

+++ backend/src/controller/upload.controller.js
@@ -1,5 +1,6 @@
 import cloudinary from "../lib/cloudinary.js"
 import fs from 'fs';
+import User from "../model/user.model.js";
 
 export const uploadDP = async (req,res) => {
     if(!req.file){
@@ -9,6 +10,14 @@ export const uploadDP = async (req,res) => {
         const result = await cloudinary.uploader.upload(req.file.path,{
             folder:'profile-pictures',
         })
+
+        const user = await User.findByIdAndUpdate(req.user._id, {
+            dp: result.secure_url
+        }, { new: true });
+        if (!user) {
+            return res.status(404).json({ message: "User not found" });
+        }
+        
         fs.unlinkSync(req.file.path);
         res.status(200).json({
             message:"Dp has been successfull uploaded",
@@ -19,5 +28,19 @@ export const uploadDP = async (req,res) => {
         res.status(500).json({message:"uploadController: Internal server error \n->"+ error});
         console.log(error);
         
+    }
+}
+
+export const removeDP = async (req, res) => {
+    try {
+        const user = await User.findByIdAndUpdate(req.user._id, {
+            dp: null
+        }, { new: true });
+        if (!user) {
+            return res.status(404).json({ message: "User not found" });
+        }
+    } catch (error) {
+        console.error("Error removing DP:", error);
+        return res.status(500).json({ message: "Internal server error" });
     }
 }
